+++ src/python/DAS/core/das_core.py
@@ -87,8 +87,6 @@ class DASCore(object):
         self.mongoparser = QLManager(dasconfig)
         dasconfig['mongoparser'] = self.mongoparser
 
-        dasroot = os.environ['DAS_ROOT']
-
         # init DAS cache
         self.rawcache = DASMongocache(dasconfig)
         dasconfig['rawcache'] = self.rawcache
@@ -97,13 +95,14 @@ class DASCore(object):
         # dasconfig; load appropriate module/class; register data
         # service with DASCore.
         self.systems = dasmapping.list_systems()
-        if  not os.environ.has_key('DAS_PYTHONROOT'):
-            msg = 'DAS_PYTHONROOT environment variable is not set'
+        if  not os.environ.has_key('DAS_PYTHONPATH'):
+            msg = 'DAS_PYTHONPATH environment variable is not set'
             raise Exception(msg)
+        dasroot = os.environ['DAS_PYTHONPATH']
         for name in self.systems:
             try:
-                klass  = '%s/DAS/services/%s/%s_service.py' \
-                    % (os.environ['DAS_PYTHONROOT'], name, name)
+                klass  = 'DAS/services/%s/%s_service.py' \
+                    % (name, name)
                 srvfile = os.path.join(dasroot, klass)
                 with file(srvfile) as srvclass:
                     for line in srvclass: